Add smart contract about solidity by example.

+++ src/attack/crosscontractorattacker/Bank.sol
@@ -2,15 +2,15 @@
 pragma solidity ^0.8.10;
 
 contract Bank {
-    mapping(address => uint) public balances;
+    mapping(address => uint256) public balances;
 
     function deposit() public payable {
         balances[msg.sender] += msg.value;
     }
 
-    function withdraw(address to, uint amount) public {
+    function withdraw(address to, uint256 amount) public {
         require(balances[to] >= amount, "Insufficient funds");
-        (bool success, ) = to.call{value: amount}("");
+        (bool success,) = to.call{value: amount}("");
         require(success, "Transfer failed");
         balances[to] -= amount;
     }
